decompose cosmetic filters
