+++ src/cosmetic_filter_cache_builder.rs
@@ -1,0 +1,247 @@
+
+use crate::cosmetic_filter_cache::ProceduralOrActionFilter;
+use crate::cosmetic_filter_utils::SpecificFilterType;
+use crate::filters::cosmetic::{
+  CosmeticFilter, CosmeticFilterMask, CosmeticFilterOperator};
+use crate::cosmetic_filter_utils::{encode_script_with_permission, key_from_selector};
+use crate::filters::fb_network::flat::fb;
+use crate::flatbuffers::containers::flat_map::FlatMapBuilder;
+use crate::flatbuffers::containers::flat_multimap::FlatMultiMapBuilder;
+
+use crate::flatbuffers::containers::flat_serialize::{
+  serialize_vec_opt, FlatBuilder, FlatSerialize,
+};
+
+use crate::utils::Hash;
+
+use std::collections::{HashMap, HashSet};
+
+
+
+use flatbuffers::WIPOffset;
+
+/// Accumulates hostname-specific rules for a single domain before building HostnameSpecificRules
+/// Note: hide and inject_script are now handled separately at the top level
+#[derive(Default)]
+struct HostnameRule {
+    unhide: Vec<String>,
+    uninject_script: Vec<String>,
+    procedural_action: Vec<String>,
+    procedural_action_exception: Vec<String>,
+}
+
+impl<'a, B: FlatBuilder<'a>> FlatSerialize<'a, B> for HostnameRule {
+    type Output = WIPOffset<fb::HostnameSpecificRules<'a>>;
+
+    fn serialize(
+        value: Self,
+        builder: &mut B,
+    ) -> flatbuffers::WIPOffset<fb::HostnameSpecificRules<'a>> {
+        let unhide = serialize_vec_opt(value.unhide, builder);
+        let uninject_script = serialize_vec_opt(value.uninject_script, builder);
+        let procedural_action = serialize_vec_opt(value.procedural_action, builder);
+        let procedural_action_exception =
+            serialize_vec_opt(value.procedural_action_exception, builder);
+
+        fb::HostnameSpecificRules::create(
+            builder.raw_builder(),
+            &fb::HostnameSpecificRulesArgs {
+                unhide,
+                uninject_script,
+                procedural_action,
+                procedural_action_exception,
+            },
+        )
+    }
+}
+
+#[derive(Default)]
+pub(crate) struct CosmeticFilterCacheBuilder {
+    simple_class_rules: HashSet<String>,
+    simple_id_rules: HashSet<String>,
+    misc_generic_selectors: HashSet<String>,
+    complex_class_rules: FlatMultiMapBuilder<String, String>,
+    complex_id_rules: FlatMultiMapBuilder<String, String>,
+
+    hostname_hide: FlatMultiMapBuilder<Hash, String>,
+    hostname_inject_script: FlatMultiMapBuilder<Hash, String>,
+
+    specific_rules: HashMap<Hash, HostnameRule>,
+}
+
+impl CosmeticFilterCacheBuilder {
+    pub fn from_rules(rules: Vec<CosmeticFilter>) -> Self {
+        let mut self_ = Self::default();
+
+        for rule in rules {
+            self_.add_filter(rule)
+        }
+
+        self_
+    }
+
+    pub fn add_filter(&mut self, rule: CosmeticFilter) {
+        if rule.has_hostname_constraint() {
+            if let Some(generic_rule) = rule.hidden_generic_rule() {
+                self.add_generic_filter(generic_rule);
+            }
+            self.store_hostname_rule(rule);
+        } else {
+            self.add_generic_filter(rule);
+        }
+    }
+
+    /// Add a filter, assuming it has already been determined to be a generic rule
+    fn add_generic_filter(&mut self, rule: CosmeticFilter) {
+        let selector = match rule.plain_css_selector() {
+            Some(s) => s.to_string(),
+            None => {
+                // Procedural cosmetic filters cannot be generic.
+                // Silently ignoring this filter.
+                return;
+            }
+        };
+
+        if selector.starts_with('.') {
+            if let Some(key) = key_from_selector(&selector) {
+                assert!(key.starts_with('.'));
+                let class = key[1..].to_string();
+                if key == selector {
+                    self.simple_class_rules.insert(class);
+                } else {
+                    self.complex_class_rules.insert(class, selector);
+                }
+            }
+        } else if selector.starts_with('#') {
+            if let Some(key) = key_from_selector(&selector) {
+                assert!(key.starts_with('#'));
+                let id = key[1..].to_string();
+                if key == selector {
+                    self.simple_id_rules.insert(id);
+                } else {
+                    self.complex_id_rules.insert(id, selector);
+                }
+            }
+        } else {
+            self.misc_generic_selectors.insert(selector);
+        }
+    }
+
+    // TODO: review this
+    fn store_hostname_rule(&mut self, rule: CosmeticFilter) {
+        use SpecificFilterType::*;
+
+        let unhide = rule.mask.contains(CosmeticFilterMask::UNHIDE);
+        let script_inject = rule.mask.contains(CosmeticFilterMask::SCRIPT_INJECT);
+
+        let kind = match (
+            script_inject,
+            rule.plain_css_selector().map(|s| s.to_string()),
+            rule.action,
+        ) {
+            (false, Some(selector), None) => Hide(selector),
+            (true, Some(selector), None) => InjectScript((selector, rule.permission)),
+            (false, selector, action) => ProceduralOrAction(
+                serde_json::to_string(&ProceduralOrActionFilter {
+                    selector: selector
+                        .map(|selector| vec![CosmeticFilterOperator::CssSelector(selector)])
+                        .unwrap_or(rule.selector),
+                    action,
+                })
+                .unwrap(),
+            ),
+            (true, _, Some(_)) => return, // script injection with action - shouldn't be possible
+            (true, None, _) => return, // script injection without plain CSS selector - shouldn't be possible
+        };
+
+        let kind = if unhide { kind.negated() } else { kind };
+
+        let tokens_to_insert = std::iter::empty()
+            .chain(rule.hostnames.unwrap_or_default())
+            .chain(rule.entities.unwrap_or_default());
+
+        tokens_to_insert.for_each(|t| self.store_hostname_filter(&t, kind.clone()));
+
+        let tokens_to_insert_negated = std::iter::empty()
+            .chain(rule.not_hostnames.unwrap_or_default())
+            .chain(rule.not_entities.unwrap_or_default());
+
+        let negated = kind.negated();
+
+        tokens_to_insert_negated.for_each(|t| self.store_hostname_filter(&t, negated.clone()));
+    }
+
+    fn store_hostname_filter(&mut self, token: &Hash, kind: SpecificFilterType) {
+        use SpecificFilterType::*;
+
+        match kind {
+            // Handle hide and inject_script at top level for better deduplication
+            Hide(s) => {
+                self.hostname_hide.insert(*token, s);
+            }
+            InjectScript((s, permission)) => {
+                let encoded_script = encode_script_with_permission(s, permission);
+                self.hostname_inject_script.insert(*token, encoded_script);
+            }
+            // Handle remaining types through HostnameRule
+            Unhide(s) => {
+                let entry = self.specific_rules.entry(*token).or_insert_with(HostnameRule::default);
+                entry.unhide.push(s);
+            }
+            UninjectScript((s, _)) => {
+                let entry = self.specific_rules.entry(*token).or_insert_with(HostnameRule::default);
+                entry.uninject_script.push(s);
+            }
+            ProceduralOrAction(s) => {
+                let entry = self.specific_rules.entry(*token).or_insert_with(HostnameRule::default);
+                entry.procedural_action.push(s);
+            }
+            ProceduralOrActionException(s) => {
+                let entry = self.specific_rules.entry(*token).or_insert_with(HostnameRule::default);
+                entry.procedural_action_exception.push(s);
+            }
+        }
+    }
+}
+
+impl<'a, B: FlatBuilder<'a>> FlatSerialize<'a, B> for CosmeticFilterCacheBuilder {
+    type Output = WIPOffset<fb::CosmeticFilters<'a>>;
+    fn serialize(value: Self, builder: &mut B) -> WIPOffset<fb::CosmeticFilters<'a>> {
+        let complex_class_rules = FlatMultiMapBuilder::finish(value.complex_class_rules, builder);
+        let complex_id_rules = FlatMultiMapBuilder::finish(value.complex_id_rules, builder);
+
+        // Handle top-level hostname hide and inject_script for better deduplication
+        let hostname_hide = FlatMultiMapBuilder::finish(value.hostname_hide, builder);
+        let hostname_inject_script =
+            FlatMultiMapBuilder::finish(value.hostname_inject_script, builder);
+
+        // Handle remaining rule types through HostnameSpecificRules
+        let hostname_specific_rules = FlatMapBuilder::finish(value.specific_rules, builder);
+
+        let simple_class_rules = Some(FlatSerialize::serialize(value.simple_class_rules, builder));
+        let simple_id_rules = Some(FlatSerialize::serialize(value.simple_id_rules, builder));
+        let misc_generic_selectors = Some(FlatSerialize::serialize(
+            value.misc_generic_selectors,
+            builder,
+        ));
+
+        fb::CosmeticFilters::create(
+            builder.raw_builder(),
+            &fb::CosmeticFiltersArgs {
+                simple_class_rules,
+                simple_id_rules,
+                misc_generic_selectors,
+                complex_class_rules_index: Some(complex_class_rules.keys),
+                complex_class_rules_values: Some(complex_class_rules.values),
+                complex_id_rules_index: Some(complex_id_rules.keys),
+                complex_id_rules_values: Some(complex_id_rules.values),
+                hostname_hide_index: Some(hostname_hide.keys),
+                hostname_hide_values: Some(hostname_hide.values),
+                hostname_inject_script_index: Some(hostname_inject_script.keys),
+                hostname_inject_script_values: Some(hostname_inject_script.values),
+                hostname_index: Some(hostname_specific_rules.keys),
+                hostname_values: Some(hostname_specific_rules.values),
+            },
+        )
+    }
+}
