+++ src/filters/mod.rs
@@ -6,4 +6,5 @@ mod network_matchers;
 pub mod cosmetic;
 pub(crate) mod fb_builder;
 pub(crate) mod fb_network;
+pub(crate) mod fb_network_builder;
 pub mod network;
