+++ src/lib.rs
@@ -20,6 +20,8 @@ pub mod blocker;
 #[cfg(feature = "content-blocking")]
 pub mod content_blocking;
 pub mod cosmetic_filter_cache;
+mod cosmetic_filter_cache_builder;
+mod cosmetic_filter_utils;
 mod engine;
 pub mod filters;
 mod flatbuffers;
